requirements.txt
cairosvg==2.7.1
pillow==10.2.0
svgwrite==1.4.3
opencv-python==4.9.0.80
python-dotenv